app/services/snowflake_service.py
import snowflake.connector
from config.settings import SNOWFLAKE_USER, SNOWFLAKE_PASSWORD, SNOWFLAKE_ACCOUNT

def get_snowflake_connection():
    try:
        conn = snowflake.connector.connect(
            user=SNOWFLAKE_USER,
            password=SNOWFLAKE_PASSWORD,
            account=SNOWFLAKE_ACCOUNT
        )
        return conn
    except Exception as e:
        print(f"Error connecting to Snowflake: {e}")
        return None