requirements.txt
boto3
snowflake-connector-python==3.14.0
requests==2.32.3
python-dotenv==1.1.0